requirements.txt
Django==1.4.3
South==0.7.6
django-admin-bootstrapped==0.3.2
wsgiref==0.1.2
django-tastypie==0.9.11
